Removed confusing use of new Array()
I argue that `[]` should be used instead of `new Array()`, and especially instead of `new Array(arg0, arg1, ...)` to avoid confusion (i.e. I was confused when I saw it).

I wasn't sure if there would be a bug or not in the case that `name` was a number and `value` was undefined. According to Firefox, it should never result in a bug. I assume Chrome behaves the same way.

https://developer.mozilla.org/en-US/docs/JavaScript/Reference/Global_Objects/Array